+++ src/cpu/x64/matmul/amx_blocking_heuristics.hpp
@@ -197,7 +197,9 @@ class bw_map_t {
 public:
     bw_map_t() = default;
 
-    float get_bw(int x) const { return linear_interpolation(multicore_bw, x); }
+    float get_bw(int x) const {
+        return linear_interpolation(multicore_bw, static_cast<float>(x));
+    }
 
     // All the following bandwidth measurements were taken on an
     // EMR machine with two NUMA domains, each containing 32 cores.
@@ -216,13 +218,13 @@ public:
 private:
     // This dictionary includes DRAM bandwidth for cores that share data.
     // The key represents the number of cores sharing, and the value is the bandwidth.
-    const std::map<int, float> multicore_bw = {
-            {32, 4.06}, {16, 3.31}, {8, 2.98}, {4, 2.39}, {2, 0.9}, {1, 2.28}};
+    const std::map<int, float> multicore_bw = {{32, 4.06f}, {16, 3.31f},
+            {8, 2.98f}, {4, 2.39f}, {2, 0.9f}, {1, 2.28f}};
 
     float linear_interpolation(
             const std::map<int, float> &points, float x) const {
         // Find the interval [x0, x1] where x0 <= x <= x1
-        auto it = points.lower_bound(x);
+        auto it = points.lower_bound(static_cast<int>(x));
         if (it == points.end()) {
             return points.rbegin()
                     ->second; // x is greater than the largest x in the map
